+++ internal/module/module.go
@@ -1,18 +1,32 @@
-package witcli
+package module
 
 import (
 	"runtime/debug"
 	"sync"
 )
 
-// Version returns the version string of this module.
+// Path returns the path of the main module.
+func Path() string {
+	build := buildInfo()
+	if build == nil {
+		return "(none)"
+	}
+	return build.Main.Path
+}
+
+// Version returns the version string of the main module.
 func Version() string {
 	return versionString()
 }
 
+var buildInfo = sync.OnceValue(func() *debug.BuildInfo {
+	build, _ := debug.ReadBuildInfo()
+	return build
+})
+
 var versionString = sync.OnceValue(func() string {
-	build, ok := debug.ReadBuildInfo()
-	if !ok {
+	build := buildInfo()
+	if build == nil {
 		return "(none)"
 	}
 	version := build.Main.Version